コマンドごとにファイルを分割
