+++ src/commands/commandUtil/SlashCommand.ts
@@ -1,0 +1,7 @@
+import discord from 'discord.js';
+
+interface SlashCommand extends discord.ChatInputApplicationCommandData {
+    execute: (client: discord.Client, interaction: discord.CommandInteraction) => Promise<void>;
+};
+
+export default SlashCommand;
