+++ src/commands/index.ts
@@ -1,0 +1,5 @@
+export * from './now';
+export * from './list';
+export * from './register';
+export * from './unregister';
+export * from './update';
